move XMLUtils and CamelMavenServiceManager to package service and reduced visibility from public to package private
