+++ camel-idea-plugin/src/main/java/org/apache/camel/idea/service/CamelCatalogService.java
@@ -22,7 +22,6 @@ import com.intellij.openapi.Disposable;
 import org.apache.camel.catalog.CamelCatalog;
 import org.apache.camel.catalog.DefaultCamelCatalog;
 import org.apache.camel.catalog.DefaultVersionManager;
-import org.apache.camel.idea.util.CamelMavenVersionManager;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -42,7 +41,7 @@ public class CamelCatalogService implements Disposable {
         return instance;
     }
 
-    public boolean isInstantiated() {
+    boolean isInstantiated() {
         return instance != null;
     }
 
@@ -52,7 +51,7 @@ public class CamelCatalogService implements Disposable {
      * @param version the version to load
      * @param repos   any third party maven repositories
      */
-    public boolean loadVersion(@NotNull String version, @NotNull Map<String, String> repos) {
+    boolean loadVersion(@NotNull String version, @NotNull Map<String, String> repos) {
         // we should load a new version of the catalog, and therefor must discard the old version
         dispose();
         // use maven to be able to load the version dynamic
