+++ camel-idea-plugin/src/test/java/org/apache/camel/idea/service/XmlUtilsTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.idea.util;
+package org.apache.camel.idea.service;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
